Improve timing in VDP::syncAtNextLine()
Several settings (screen mode, display-enable, ...) are fixed per line.
That is, the V99x8 looks at the settings at a specific moment during
(line-)rendering, and then renders a full line using those settings. In
other words: you cannot switch screen mode or enable/disable rendering
in the middle of a line. Note: this only applies to specific settings
(stuff like R#23 vertical scroll can be changed in the middle of a
line).

This patch fixes the specific moment within the line where the settings
are latched. This is implemented in the VDP::syncAtNextLine() method.

All credits to grauw for:
- Reverse engineering the 'Psycho World' Z80 code (related to the
  screen-split stuff).
- Testing multiple hypothesis on what could cause wrong emulation in
  openMSX. Done by creating (Z80) cycle accurate test programs and
  running them on real MSX machines.
- Follow-up his initial tests with more (related) tests. In other words:
  fix the problem in general, not just one specific manifestation.
- Writing a very clean openMSX patch.

See here for more details:
    https://github.com/openMSX/openMSX/issues/1310