Fix parsing of 'color_matrix' setting
Revision 469f335d59f replaced manual matrix/vector calculations with
matrix/vector-type calculation. This greatly simplifies the code

For example:
    float outR = m[0][0] * inR + m[0][1] * inG + m[0][2] * inB;
    float outG = m[1][0] * inR + m[1][1] * inG + m[1][2] * inB;
    float outB = m[2][0] * inR + m[2][1] * inG + m[2][2] * inB;
becomes:
    vec3 outRGB = m * inRGB;

Unfortunately there was a mismatch between column/row-major order before
after this transformation. In other words it's as-if the color-matrix
was transposed after this patch. (Un)fortunately many of our example
color matrices are symmetrical or near symmetrical. So for many (but not
all) the visible difference was minimal.

Fixed by adjusting the color-matrix parsing code to fill in the
rows/columns using the correct convention.

This fixes:
    Bug#1312 "monitor_type amber" command not working properly
    https://github.com/openMSX/openMSX/issues/1312