Added tmpStrCat() as a (faster) alternative to strCat()
We have a lot of code that follows this pattern:
    void CassettePlayer::TapeCommand::execute( ..., TclObject& result) {
        ...
        result = strCat("Motor control is ", (motorControl ? "on" : "off"));
        ...
    }
So we concatenate short string fragments into a larger string and assign
that to a TclObject.

Another pattern is to concatenate directory, filename, file-extension
into a larger string and pass that to some 'FileContext' resolve()
function. (And still several other patterns).

The common theme is that we first create a std::string object (for all
but the shortest strings (typically 15 characters or less) this involves
heap allocation). Then call a function which doesn't really require a
std::string (a std::string_view is sufficient). And when that function
returns we free that std::string. In other words it's a short lived
temporary string. Tools like 'heaptrack' flag the associated heap
allocations as 'temporary allocations'.

'Temporary allocations' are often easy to eliminate (e.g. move from heap
to stack). And that's what this patch does. It adds a new function
'tmpStrCat()' as an alternative for 'strCat()'. The only difference is
that it returns a 'TemporaryString' instead of a 'std::string'.
'TemporaryString' is also introduced in this patch. It's more or less
like a 'std::string' but with a much larger internal buffer, so it can
store larger strings before requiring heap allocation. In general
there's a good reason why the buffer in std::string shouldn't be too
large. But in specific cases (like the examples above) other trade-offs
can sometimes be better.

In practice the 'TemporaryString' buffer is large enough to hold most
(all?) of our short lived strCat() results. I did a first pass of
replacing strCat() with tmpStrCat() and that already gave almost 100
less allocations during startup. And there's potential for more saving
in the future.

It's also worth noting that replacing 'strCat()' with 'tmpStrCat()' is
safe. In the sense that the code won't compile when the transformation
isn't valid (e.g. when the called code does expect a std::string).

---

This patch introduces a new function and then the programmer has to
choose between 'strCat()' and 'tmpStrCat()'. My initial attempt was to
keep a single 'strCat()' function and then transparently, depending on
the context, create either a 'std::string' or a 'TemporaryString'.
(Implementation detail: strCat() returns an object that only captures
the parameters, and that object has conversion operators to std::string
and std::string_view). This worked fine 90% of the time, but the
remaining 10% (mostly related to templated functions) was annoying
enough that in the end I decided against this approach.