Fix #1104: disabling sprites should not be instant
Like screen-mode and display-enabled (see rev 547064a2e6) the
sprite-enabled setting is only latched once per line.

I already wrote a patch for this in May 2018, see attachment in
    https://github.com/openMSX/openMSX/issues/1104
Though because of a bug in that patch (initialize 'spriteEnabled' to
'false' instead of 'true') it caused a hang on turboR machines and
therefor the patch never got applied.

Thanks a lot to 'grauw' for debugging that patch!