Fix(??) msvc compile error
Code compiled fine with gcc/clang but not with msvc. Not sure why
exactly, not even sure it's a bug in msvc.

(Wild guess: it might be because gcc/clang perform copy-elision and msvc
does it. According to the c++ standard this is an allowed but not a
required optimization. But shouldn't gcc/clang then still complain about
a delete move operation (even if it is optimized away) like msvc does
now?)

Anyway, I did some experiments on godbolt, and this patch _might_ fix
it. It's in any case more general.