Added 'zstring_view'
Consider this code fragment:
    string text(tokens[1].getString());
    SDL_SetClipboardText(text.c_str());

Here 'tokens[1]' is a TclObject. The getString() method returns a
std::string_view, not a std::string (this avoids making a copy of the
internal TclObject string-representation). But we want to pass this to
the SDL_SetClipboardText() function which expects a zero-terminated
c-string. 'std::string_view' does not guarantee that it refers to a
zero-terminated string. That's why the current code first needs to copy
the std::string_view to a std::string (because std::string does offer
this guarantee).

But actually the original internal string representation in TclObject
_does_ already guarantee zero-terminated strings. It's only because we
return a std::string_view (to avoid making a copy) that we loose this
guarantee (and then are required to make a copy later).

This patch introduce the 'zstring_view' class. Like 'string_view' it
only refers to a string (it does not own the data, so it does not
require making a copy). But it has the additional guarantee that it
refers to a zero-terminated string (because of this it e.g. cannot refer
to a part in the middle of a larger string). This allows to fully
eliminate making copies in the above example.

More in general using zstring_view can be useful when interacting with
3rd-party APIs which expect C-style strings (so also e.g. file operations
in libc). You could ask why not simply use 'const char*' for such
situations. Often that's possible, but:
- This looses the length information, (this info is available for free,
  and loosing it is sometimes the source of other performance
  inefficiencies).
- 'zstring_view' carries more semantic information than 'const char*'.
  In some cases this helps to detect programming errors.