Extend 'user_setting' command with 'enum'
The 'user_setting' command allows to create extra openMSX settings from
Tcl scripts (in addition to those defined in the C++ code). Before this
patch those settings were limited to 'boolean', 'integer', 'float' and
'string' settings. This patch extends that list with 'enum' settings.
(This is sort of a 'string' setting, but with a fixed list of allowed
values).

This patch only introduces the possibility. The next patch will actually
use this new feature.