Add 'monitor_type' setting
We have the 'color_matrix' settings which defines a general 3x3 matrix
to transform the MSX-RGB colors to host-RGB colors. Though this is
pretty low level, and difficult to use. Therefor we added the
'monitor_type' Tcl proc which allows to easily switch between several
predefined matrices. (E.g. try 'monitor_type monochrome_amber').

We recently got a request to include 'monitor_type' in the OSD menu.
That's a good idea, but (because of implementation details in the menu
code) it would be much easier to implement if 'monitor_type' were a
setting instead of a 'proc'. A setting (instead of a proc) would also be
a natural fit for this functionality. I didn't look it up, but probably
at the time we introduced 'monitor_type' we didn't have user defined
settings yet, and certainly no user defined enum settings.

Anyway, this patch adds such an enum setting. So we now have both a
'monitor_type' proc and a 'monitor_type' setting. I kept the proc, both
for backwards compatibility and also because we anyway need such a proc
to translate the enum-value to the corresponding 'color_matrix' value.
In other words: without backwards compatibility this could be an
internal helper proc, now it's exposed. That's OK.

This patch does not yet extend the OSD menu.