Tweak some uses of FileOperations::join()
'FileOperations::join()' is a great general purpose way to join two or
more (unknown) filepaths (e.g. the 2nd path might be an absolute path
and then the 1st path gets discarded). But in specific cases we
sometimes know more details about the paths (e.g. we know the 2nd path
doesn't start with '/') and then we can do better. Here's one example:

    FileOperations::join(type, tmpStrCat(name, ".xml"));

'type' is the string "extensions" or "machines" and 'name' is the name
of the machine or extension. So we first append the ".xml" extension and
then join with the appropriate directory name. This makes 2 intermediate
copies and (typically) allocates 1 temporary buffer (only 1 instead of 2
because of tmpStrCat()). If instead we write:

    tmpStrCat(type, '/', name, ".xml");

Then there's only 1 copy of 'name' and no memory allocation. IMHO the
2nd version is also not less readable than the first one.