Added StringOp::split_view() (+unittest) ..
.. and use it to tweak text manipulation in OSDText.

StringOp::split() takes a string and a delimiter character and returns a
vector of sub-strings. StringOp::split_view() does the same, but instead
of eagerly doing all the splits, it does it in a lazy way. That is it
splits the first sub-string, then lets the caller process that string,
then splits the seconds sub-string and so on. This avoids having to
allocated memory (a vector) to hold the collection of all sub-strings.