Use 'perfect forwarding' in XMLElement API
'Perfect forwarding' is a technical term, described here:
    https://stackoverflow.com/questions/6829241/perfect-forwarding-whats-it-all-about
    https://en.cppreference.com/w/cpp/utility/forward
In short: it's a c++ template technique that allows to forward both the
lvalue- and rvalue-ness from a (templated) function parameter to another
function.

More concrete, in this patch we use it to further optimize
string-construction/assignment in the XMLElement class.

We already optimized XMLElement to take (sink) string parameters by
value, and then move those parameter to the member variable. For example

    void XMLElement::setName(std::string newName) {
        name = std::move(newName);
    }

This avoids excess string copies. This makes at most 1 string copy (on
the caller side), and only when really needed. (E.g. it makes no copy at
all when passed an r-value string. And only 1 string-construction (not
construction + copy) when passed a string-literal). In the callee
there's exactly 1 string-move. Moving a std::string is cheap, but not
completely free either.

A disadvantage is that the string object needs to be constructed at the
caller side. Often this is done transparently but not always. For
example:

    // passing string-literal is transparent
    someElement.setName("foobar");
    // but passing std::string_view requires explicit conversion
    someElement.setName(std::string(someTclObject.getString()));

It would be nice if we didn't need to clutter (some) callers with this.

Another disadvantage is that this cannot reuse existing capacity. For
example when the 'name' member variable already contains a string with
capacity for N characters and we assign a string-literal of length <= N,
then we could just copy over those characters (no need to construct a
new string for that literal and move it over the existing string).

This patch addresses all the above problems. The setName()
implementation now looks like:

    template<typename String>
    void XMLElement::setName(String&& newName) {
        name = std::forward<String>(newName);
    }

I'm skipping over the technical details, but in short: this method now
accepts any (string-like) parameter-type and 'perfectly forwards' it to
any of the overloaded string::operator= methods. I'm simplifying, but
std::string has these 6 overloads:

    1) string& operator=(const string& str);
    2) string& operator=(string&& str);
    3) string& operator=(const CharT* s);
    4) string& operator=(CharT ch);
    5) string& operator=(std::initializer_list<CharT> ilist);
    6) string& operator=(string_view v);

Before this patch we always used the 2nd overload. Now we forward to the
most appropriate overload. As a bonus this also avoids the (cheap but
not free) string-move operation.

In summary, this patch makes a trade-off between:
- More complex implementation in the callee (template instead of a
  regular method).
- Possibly simpler code in (all) callers.
- (Bonus) in some cases more efficient.

In case of XMLElement, we have many callers (and only one implementation
of XMLElement). So this trade-off seems worthwhile. (I also measured
almost 100 fewer memory allocations during openMSX startup).