Hold MSXMotherBoard objects by shared_ptr in Reactor
This fixes a possible crash in Tcl callbacks that switch the active
machine in the middle of a CPU instruction.

The following scenario (discovered by Manuel) triggered the problem:
    Write this in the console:
        debug set_watchpoint write_mem 0x9000 {} {step_back}
    Then in basic do:
        POKE &H9000,1

The Tcl callback triggers while the CPU instruction that does the memory
store is still being emulated. In the callback we execute the
'step_back' script. This executes 'reverse goto' command(s). Internally
those commands are implemented by creating a new MSX machine instance,
emulating that instance to the requested moment in time, and then
switching from the current to that new MSX machine instance. After that
the old machine instance gets deleted. After the callback finishes, the
store instruction continues emulating, but that's now operating on an
already deleted MSX instance. And that leads to the crash.

We already had similar problems in the past. And to fix those we
introduced the 'garbageBoards' mechanism in Reactor. This is a primitive
form of 'garbage collection'. That is instead of deleting the old MSX
instance immediately after switching, we postpone this till a 'safe'
moment in time. More specifically we delete in the main event loop
(where we know that no machines are 'actively' being emulated, IOW there
are no active stack-frames that have a reference to the machine).

The situation got more complex after commit '3054bdcb07', see that
commit message for a much more detailed description. Short summary: in
specific cases the 'step_back' script has to execute a lot of internal
steps. Each step creates a temporary machine. In garbage collector
terminology: this creates a lot of 'garbage' before the 'collector' has
a chance to run. So while functionally 100% correct, it caused a large
memory usage spike, and in specific scenarios this could cause problems.
(~8 year ago when a typical desktop machine had less memory than today).

To fix this memory spike problem we bypassed the 'garbageBoards'
mechanism in specific cases (when switching boards via 'reverse'). It's
fine to already delete the machine from a Tcl script as long as when the
script finished, all the active stack-frames return without still
accessing the machine state. And we audited/rewrote code to ensure this.
However I overlooked the possibility that Tcl scripts can also trigger
in the middle of a CPU instruction.

To (hopefully?) fully fix this, this patch replaces the primitive
'garbageBoards' mechanism with a more invasive 'reference counting'
mechanism (implemented using 'std::shared_ptr'). That is each machine
instance gets an associated reference count. When a machine is actively
being emulated that count gets temporarily increased, to ensure the
machine does not get deleted too soon. Conversely as soon as the
reference count of (another) machine drops to zero we can already delete
it, no need to wait till the main event loop runs.

So far all tests went well. All previously known scenarios seem to work
fine, without memory usage spikes. But still can use more testing.