Fixed internal error in 'step_back' Tcl script
This is a followup on a23e15964c157bc. The motivating use case for that
patch was:
    Write this in the console:
        debug set_watchpoint write_mem 0x9000 {} {step_back}
    Then in basic do:
        POKE &H9000,1

After a23e15964c157bc this no longer crashes, but it still produces the
following Tcl error:
    Internal error: overshot destination

The reason for this was that the 'step_back' script did not expect to be
called from somewhere (int time) in the middle of a CPU instruction.
After this fix the error is gone, and the above use case goes in an
infinite loop of always re-executing the same store instruction (as it
should).

Some background info:

We don't keep track of the length of emulated CPU instructions. This
would slow down normal emulation. But if we don't have this info, how
can we go back to (the moment in time of the start of) the previous
instruction? The trick is to go back a 'large' step (the length of the
longest possible CPU instruction) and from there on take single
instruction forward steps. We expect to end up at exactly the starting
point again. During this single-step phase we do track the start time of
the previous instruction. Once we reach the starting point we can take
one final backwards step again. See comments in the script for more
details.

The above assumption that we end up at exactly the starting point must
be relaxed when step_back can be called from some time point in the
middle of an instruction.