Fix tab-completion in user-directories
For example tab-completion on
    reverse loadstate r<tab>
should complete all filenames (that start with 'r') in both the current
directory and the '~/.openMSX/replays' directory. The latter were
missing after revision cab2f971bc7af9 (2020/08/20).

Problem was that internally the possible completions were matched
against the full absolute path instead of the relative portion of the
path (so e.g. the '/home/wouter/.openMSX/replays' prefix should be
dropped).