[1/5]: Use exceptions in SectorAccessibleDisk::readSectors()
Use exception instead of returning an error code in
SectorAccessibleDisk::readSectors() and writeSectors(). This is more
inline with the rest of the openMSX code.

The original code was like this to be more similar to external Nowind
code. Though that code hasn't changed in +10 years. And this change in
API is small enough that it still shouldn't be a problem if we ever want
to integrate new Nowind code.

Later patches in this series will need to read multiple sectors in one
call. I thought it was easier to reuse the existing readSectors() method
than to introduce a new one.