[2/5] Added SectorAccessibleDisk::readSectorsImpl()
Before this patch:
- SectorAccessibleDisk::readSectors() simply had a loop that called
  readSector().
- readSector() (via IPS patching mechanism) calls readSectorImpl().
- All SectorAccessibleDisk subclasses override readSectorImpl() to
  handle reading a single sector.

After this patch:
- readSector(sector, buf) calls readSectors(&buf, sector, 1)
- readSectors() (via IPS patching mechanism) calls readSectorsImpl()
- readSectorsImpl() is a new method whose default implementation calls
  readSectorImpl() in a loop.

So basically nothing changes. Calling either readSector() or
readSectors() still results in (one or more) calls to readSectorImpl().

The point of this patch is that we now go via the new (virtual)
readSectorsImpl() method. And that we have the _possibility_ to override
that method. In the next patch in this series we'll exploit this
possibility.

ATM we only do this for reading sectors. Nothing prevents us from doing
the same for writing, but ATM it is not really needed yet. That is:
future patches in this series can't take advantage of writing multiple
sectors at once.