[3/5] Override readSectorsImpl() for (some) subclasses
Override readSectorsImpl() instead of readSectorImpl() for some
subclasses of SectorAccessibleDisk.

This can be more efficient when reading more than 1 sector at once. For
example it can do one file-read system call or one memcpy() rather than
a loop which reads or copies one sector (512 bytes) at a time.

This patch enables the more efficient read, but there's no code in
openMSX yet that exploits it. That is openMSX still calls readSector()
instead of readSectors(). The next patch in this series will change
that.