[5/5] Increase TigerTree block-size from 1kB to 64kB
We use TigerTree to incrementally calculate a hash for a HD image. That
is if only one sector on the HD changes, we don't have to redo the full
hash calculation. Instead we only have to recalculate the changed block,
and then propagate this result up a binary-tree.

By changing to a larger block-size, we make a different trade-off
between:
- Larger hash (re-)calculation at leaf-nodes.
- Shallower tree-depth.

In addition, in the current implementation, the block-size determines
the number of calls to readSectors(). Which directly translates to the
number of read-file system calls we make.

Combined with the previous patches in this series, this reduces the
number of read-file system calls by a factor 128x.

For me, with a 100MB harddisk image, on Linux on a local ext4
filesystem, this improved startup time from 0.40 to 0.14 seconds. But
the results may be very different (larger improvement?) on Windows on a
SMB network filesystem. And that was in fact the whole motivation for
this patch series. See the discussion in:
    https://www.msx.org/forum/msx-talk/openmsx/ide-hdd-via-smb