Fix (theoretical) undefined behavior in call to memcpy()
Strictly according to the c++ standard it's undefined behavior to call
    void *memcpy(void* dest, const void* src, size_t n);
when either 'dest' or 'src' is 'nullptr', even when 'n == 0'.

However no practical implementation of memcpy() will actually
dereference 'dest' nor 'src' when 'n == 0'. So in practice this will
work just fine. There even is talk about relaxing the standard on this
point.

In openMSX we triggered this "undefined behavior" when using strCat()
and one of the components was an empty std::string_view.

When we also compile openMSX with "-fsanitizer=undefined" we get an
error pointing out this corner case. I'm trying to get the number of
these errors to zero, so let's fix this (non-)issue.