Change implementation-define-behavior to standard code
In our sha1 code we had some SSE2 optimized code which did:

    // load 64-bit (possibly unaligned) and swap bytes
    [[nodiscard]] static inline uint64_t loadSwap64(const char* s)
    {
        return Endian::byteswap64(*reinterpret_cast<const uint64_t*>(s));
    }

Casting a 'char*' to 'uint64_t*' and reading from it is undefined
behavior in case the pointers are not 64-bit aligned. Actually it's
implementation-defined behavior, and because this is used in the context
of SSE2 code, we know we're running on an x86 platform, and this code is
perfectly fine.

The only 'problem' is that the undefined-behavior-sanitizer
(-fsanitize=undefined) does flag it. (And we try to get to zero errors).

Though it's very easy to fix. In fact we already had a fully standards
conforming version of this function. So using it makes the code also
simpler. And I verified that the compiler generates identical code for
both.