Use named invalid-scaler enum instead of casting
Because casting an arbitrary integer to an enum (which is not one of the
numeric values used by the enum) is undefined behavior. (But often just
works in practice).

But we already had a named enum value that represented the state "no
valid scaler". So doing this in a standard compliant way also cleans up
the code.