Switch from '0' to '-1' to indicate an invalid index
Before this patch, in hash_set_impl::Pool we used the value '0' to
indicate an invalid index. This patch switches that to '-1' (actually we
use the constant 'invalidIndex' with numeric value '-1').

I'm not exactly sure why I originally picked '0'. Maybe because it makes
some code shorter?
    if (!index) { ...
vs
    if (index == invalidIndex) { ...
Or maybe testing for '0' is a tiny bit more efficient than testing for
'-1'? (In some situations on some platforms). Anyway, I think this
difference will be neglible.

Though a consequence of this choose was that valid indices started at
'1' instead of '0'. But we want to be able to use the element at
position '0' in our buffer. To compensate for that we subtract 1 from
the pointer to our internal buffer. This way we only have to deal with
this offset once, when the buffer get allocated, instead of on each
index operation into the buffer.

However strictly according to the c++ standard this is undefined
behavior. It's not allowed to form a pointer that points outside the
bounds of the array (one past the array is allowed, one before the array
is not). Note that before dereferencing the pointer we always add an
offset (at least 1 in this case) so that the result IS again inside the
array bounds. Nevertheless it's wrong: just forming the pointer is
already undefined behavior. In practice, on most platforms, there's no
problem at all. But there is some exotic hardware (typically DSPs) that
can misbehave.

This form of undefined behavior is caught by UBSAN (compile with
-fsanitize=undef). I'd like to get to zero ubsan warnings, so let's fix
this. Luckily this isn't too difficult.

In this patch I've also switched from directly using the type 'uint32_t'
to the type 'PoolIndex' (which is just a struct containing a single
'uint32_t'). This stronger typing helped make sure I did the conversion
correctly (e.g. the old implicit test for zero no longer compiles).