Fix ubsan errors in combination with VLA
The undefined behavior sanitizer caught two problems related to usage of
VLA (variable length array).

1) VLAs of size zero are not allowed. Though gcc/clang support it as an
extension. (Actually VLA in c++ is already an extension). Easily fixed.

2) VLA bound-checks when that VLA is captured by a lambda expression
seem to go wrong. I think this is a limitation in the ubsan
implementation. Though strictly speaking VLAs should not be supported in
c++ at all. Luckily in this case the VLA could easily be replaced with a
fixed size array (with a very reasonable upper bound).

--

After this patch I can't trigger any ubsan errors anymore (in a
reasonable set of manual tests).