Math::findFirstSet() cannot (yet) be constexpr on msvc
Because the msvc function _BitScanForward() is not constexpr. Apparently
on gcc/clang the corresponding function __builtin_ffs() IS constexpr.

Could be fixed by using c++20 std::is_constant_evaluated(). But we
cannot use that yet, and for now we also don't really need a constexpr
Math::findFirstSet(). So just revert this (small) part of the previous
patch.