Don't use deprecated std::shared_ptr<T>::unique()
The functions:
    std::shared_ptr<T>::use_count()
    std::shared_ptr<T>::unique()
respectively query the use-count of a shared_ptr or test if that count
is equal to 1. These are deprecated in c++20 and msvc already gives
useful warnings for this. The reason for the deprecation is that it's
unsafe to use these functions from multiple threads (while all other
shared_ptr functionality is thread safe). See here for more details:
    https://stackoverflow.com/questions/41142315/why-is-stdshared-ptrunique-deprecated

We did use std::shared_ptr<T>::unique(). We only called it from a single
thread, so it should have been safe. Though it's easy enough to rewrite
the code to not use std::shared_ptr, but instead add an (intrusive)
use-count in the 'Decompressed' class. (This may also be more
performant, but not important in this context).

The reference counting mechanism is used to avoid having to decompress
the exact same file multiple times. E.g. when you use a machine with
compressed system roms or insert a compressed disk image, and then
reverse we (temporarily) create a new machine instance with the same
system roms and same disk image.

This reference counting mechanism also allows to share the decompressed
memory buffer between the different machine instances (to save some
memory). Though while working on this patch I noticed that the shared
memory buffer was writable. Luckily no code actually wrote to the
buffer, but it's safer to enforce read-only sharing in the code. So this
patch makes the File::mmap() function return 'span<const uint8_t>' (was
'span<uint8_t>').