Make more functions constexpr
I don't believe many of the converted functions will actually be
evaluated at compile time. But constexpr also acts as a good hint(*)
that it's a mathematically pure function. E.g. the function cannot
depend on stuff that can only be known during run-time (like global
variables).

(*) It's only a good indication (hint) rather than a 100% guarantee,
because it's fine if there is at least 1 control-flow-path through the
function that could be evaluated at compile. And (I'm not 100% sure
about this) I think compilers are not _required_ to diagnose this until
you actually try to evaluate the function at compile time (clang seems
to do more upfront checks than gcc).