In AmdFlash, replace vector<> with span<> constructor parameter
This patch changes the type of the vector<SectorInfo> constructor
parameter to span<SectorInfo>. This parameter determines the size and
the flash rom, how it's subdivides in sectors (=units that must be
erased at once) and whether these sectors are made write-protected.

Typically a specific MSX device uses a specific type of flash rom, and
thus the flash rom configuration is statically known. There's only one
exception: NowindInterface. That one does support flash roms in a
variety of size (not sure that's accurate wrt the real hardware).

Anyway, because this configuration is (mostly) known at compile-time, we
could create a compile-time data-structure (std::array) for it, rather
than a run-time data-structure (std::vector).

So this patch changes the constructor parameter from vector to span.
This makes AmdFlash agnostic to the actual underlying datastructor that
was passed (as long as it's a continuous memory block it's fine). And
that allows to use std::array for most callers, and std::vector only for
NowindInterface.