Change return type from 'const string&' to 'string_view'
This patch changes the return type for (some) virtual methods that
return 'const std::string&' to 'std::string_view'.

This can be more efficient when (some of) the overrides return a
string-literal rather than (a reference to) some member variable. This
is because in order to correctly return a literal you can't just return
that literal (that would create a reference to a (deleted) temporary).
Instead we used to create a 'static const string' for that literal and
return that.

Often (but not always) the callers of these virtual methods are just as
happy with a std::string_view. And those can be efficiently created for
both string literals and for std::string member variables.

If all of the implementation return a member variable (e.g. maybe the
method is non-virtual, and thus there's only one implementation). Then
there's no advantage in changing the return type.