Change vector -> array in tabCompletion()
Tab-completion on e.g. the 'carta' command allows for two extra strings
'eject' and 'insert' for the 2nd token. We used to dynamically allocate
a vector with these two string. But this vector has a fixed size and
content, so we could as well use a (constexpr) std::array for it.

Also prefer to use std::string_view over string-literals (const char*)
because the former captures the string length (at compile-time) and the
latter calculates the length via a call to strlen() (at run-time).