Cleanup mapper configuration in MSXSCCPlusCart
Before this class contained (among others) these 3 members:
    /*const*/ byte mapperMask;
    /*const*/ bool lowRAM, highRAM;

The commented out 'const' keyword meant that these are logically const,
but for technical reason can't be. More in detail: these members don't
change after the object is constructed. But they are not set in the
constructor's member initializer list, instead they only get initialized
in the constructor's body. This patch restructures the code so that
these can be actually 'const'.

In the original code the members 'lowRAM' and 'highRAM' made sense. But
in revision 3269b61feb1 the mapper got generalized with the 'Popolon'
subtype, but then the meaning of 'low/highRAM' became a bit murky. This
patch replaces the above three members with this struct:
    struct MapperConfig {
        unsigned numBlocks;  // number of 8kB blocks
        byte registerMask;   // mapper selection bits to ignore
        byte registerOffset; // first mapped block
    };
This can represent all the required mapper configuration. And IMHO also
in a more direct/cleaner way.

TODO revision 3269b61feb1 added the 'Popolon' subtype, and implicitly
assumed mirroring when e.g. block 255 is selected when only 1024kB RAM
is present (=128 blocks). I added a comment about this assumption. But
we should verify this on real hardware.