Fix 'cursor right' on the new 'OSD icon set' menu item
The problem was that 'cycle icons' did not advance to the next item,
while 'cycle_back icons' worked fine (hence cursor-left worked but
cursor-right did not).

This problem was in turn caused by duplicates in the list of possible
enum-values for the 'icons' setting. In my installation this list looked
like:
   none none set1 set1 set2 set2 ...
'cycle' is implemented as: search for the current value in the list, add
one to that index and set the value as that position. So if every item
is repeated this indeed explain why 'cycle' doesn't work (and also why
'cycle_back' does work: it subtracts one from the position).

The reason why I had every value twice in the list is that the list is
constructed by searching for possible sets in both the user and the
system directory. Apparently I had (the same sets) installed in both
locations.

Until now this wasn't a problem because this list was only used for
tab-completion, and there duplicates don't matter. Now we also use the
list as possible values for an enum setting.

Anyway, easy to fix with a Tcl "lsort -unique".

This patch also factors out the helper proc 'all_icons'.