Mark the 'Observer::update()' method (and overrides) 'noexcept'.
Manuel reported a crash (see next patch for more details). It took me
quite some time to figure out what was going on. But in the end I found
it was triggered by an Observer::update() callback that throws an
exception.

Subject::notify() calls the Observer::update() callbacks in a loop. But
it is unable to deal with throwing callbacks. It's also not easy (and
probably also not a good idea) to extend it so that it can deal with
them. So Implicity the code already assumed that update() callbacks
don't throw.

This patch documents that assumtion by marking the update() methods
'noexcept'. 'noexcept' is a (new) keyword in c++11. It's a promise that
a function will not throw any exception. However this is not checked at
compile time, instead if the function were to throw an exception the
program will terminate. (The compiler can in some cases generate simpler
code at the caller-side when calling a noexcept-function).

This change would have made debugging the bug much easier: now you
immediately see in the debugger which update() function (indirectly)
throwed the exception.

This patch only adds the 'noexcept' annotations. The next patch will
actually fix the bug.