Fix possible crash on Tcl error in load_icons.tcl
Discovered by Manuel: when there's an error in the
'share/scripts/load_icons.tcl' script (e.g. add "error bla" at the 2nd
to last line). Then this could result in a crash of openMSX itself!

I'm skipping over some details, but the problem was that the script
installs traces on various settings (e.g. on the 'power' setting to
monitor the status of the power LED). But unless the script runs to
completion, when these traces trigger they cause Tcl errors in the trace
proc (because they depend on Tcl variables that are not yet defined
until the end of the script).

Later in the openMSX c++ code, the 'power' setting gets changed. But
because of the trace this triggers Tcl code and in the above scenario
that Tcl code triggers an error which gets reported as a exception to
the c++ code that tried to change the value of the 'power' setting.

This happened in a c++ Observer::update() method which are not supposed
to propagate exception (see previous commit). To fix that, this patch
blocks the exception from propagating out of the method.