Use existing setting and make use of user enum settings
There was already a user setting with the icon set: osd_leds_set. But it was a
string setting. This patch removes the previously added 'icons' setting and
changes the existing string setting into an enum setting. As we're there
anyway, I also converted the osd_leds_pos string setting to enum setting.