Make more (virtual) methods 'noexcept'
This is a follow-up on 29ac5a77fb0899 which made Observer::update()
'noexcept'. But next to the Subject/Observer-mechanism addressed in that
patch, we have other similar mechanism where some object can subscribe
to events/notifications of some other object. Examples are:
    EventDistributor + EventListener
    MSXEventDistributor + MSXEventListener
    StateChangeDistributor + StateChangeListener
    GlobalCliComm + CliListener
    Display + VideoSystemChangeListener
    Layer + LayerListener

I audited the code and (i think) most listeners are not supposed to
throw exceptions. I think currently they don't, but if they did (after a
future code change) then that would be a bug. The 'noexcept' annotation
should help catch those bugs earlier.

One notable exception is 'StateChangeListener' those _can_ throw
exceptions. For example when a 'MSXCommandEvent with content "diska
non-existing-file.dsk' gets distributed (either in direct-mode or during
replay). I checked, and in this case the code is prepared to deal with
such exceptions.