Fix compilation with gcc-8
Compilation with gcc-8 was broken because we (unintentionally) already
depended on a constexpr all_of() function.

Also fixed a (false-positive) compiler warning with gcc-8.

The intention is that openMSX compiles fine with a conforming c++-17
compiler. Unfortunately that rules out gcc-7 because it misses the
<charconv> header (for the to_chars() and from_char() function).
Therefor I increased the required gcc-version from gcc-7 to gcc-8 in the
compilation manual.

Thanks to Cristi Mitrana for reporting these issues.